+++ google/analytics/client/src/main/kotlin/org/http4k/connect/google/analytics/action/GoogleAnalyticsAction.kt
@@ -1,12 +1,9 @@
 package org.http4k.connect.google.analytics.action
 
 import dev.forkhandles.result4k.Result
+import org.http4k.connect.Action
+import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.RemoteFailure
-import org.http4k.connect.google.model.TrackingId
-import org.http4k.core.Request
-import org.http4k.core.Response
 
-interface GoogleAnalyticsAction<R> {
-    fun toRequest(trackingId: TrackingId): Request
-    fun toResult(response: Response): Result<R, RemoteFailure>
-}
+@Http4kConnectAction
+interface GoogleAnalyticsAction<R> : Action<Result<R, RemoteFailure>>